requirements.txt
tweepy
numpy
scipy
Pillow==2.6.0
matplotlib
